Renamed neutral to normal, as it seems to be more commonly use (normal vision).
